IR: Separate visitor and transformer tests and clean imports
